Aktualizovat main.blocks, main.ts

+++ main.ts
@@ -2,7 +2,12 @@ input.onButtonPressed(Button.A, function () {
     Střel.move(-1)
 })
 input.onButtonPressed(Button.AB, function () {
-    sprite = 0
+    if (Postava == 0) {
+        Start = 0
+        sprite = 0
+    } else {
+        control.reset()
+    }
 })
 input.onButtonPressed(Button.B, function () {
     Střel.move(1)
@@ -13,21 +18,30 @@ function Vytvor () {
     Loď.turn(Direction.Left, 90)
 }
 input.onLogoEvent(TouchButtonEvent.Pressed, function () {
-    sprite = 1
-    Loď.delete()
-    Střel.delete()
-    game.gameOver()
+    if (sprite == 0) {
+        sprite = 1
+        Loď.delete()
+        Střel.delete()
+        game.gameOver()
+    }
 })
 let Loď: game.LedSprite = null
 let Střel: game.LedSprite = null
 let sprite = 0
+let Postava = 0
+let Start = 0
+Start = 1
+Postava = 0
 sprite = 1
-game.setScore(0)
+game.setScore(1)
 Vytvor()
 basic.forever(function () {
-    while (sprite == 0) {
-        Loď.move(-1)
-        basic.pause(500)
+    if (Start == 0) {
+        Postava = 1
+        while (sprite == 0) {
+            Loď.move(-1)
+            basic.pause(500)
+        }
     }
 })
 basic.forever(function () {
@@ -46,5 +60,8 @@ basic.forever(function () {
         game.addScore(-1)
         basic.showString("" + (game.score()))
         basic.pause(700)
+    }
+    if (game.score() == 0) {
+        game.gameOver()
     }
 })
